+++ test/session.js
@@ -8,7 +8,9 @@ var http = require('http')
 var https = require('https')
 var request = require('supertest')
 var session = require('../')
-var util = require('util')
+var SmartStore = require('./support/smart-store')
+var SyncStore = require('./support/sync-store')
+var utils = require('./support/utils')
 
 var Cookie = require('../session/cookie')
 
@@ -609,6 +611,31 @@ describe('session()', function(){
     })
   })
 
+  describe('when session without cookie property in store', function () {
+    it('should pass error from inflate', function (done) {
+      var count = 0
+      var store = new session.MemoryStore()
+      var server = createServer({ store: store }, function (req, res) {
+        req.session.num = req.session.num || ++count
+        res.end('session ' + req.session.num)
+      })
+
+      request(server)
+      .get('/')
+      .expect(shouldSetCookie('connect.sid'))
+      .expect(200, 'session 1', function (err, res) {
+        if (err) return done(err)
+        store.set(sid(res), { foo: 'bar' }, function (err) {
+          if (err) return done(err)
+          request(server)
+          .get('/')
+          .set('Cookie', cookie(res))
+          .expect(500, /Cannot read property/, done)
+        })
+      })
+    })
+  })
+
   describe('proxy option', function(){
     describe('when enabled', function(){
       var server
@@ -861,6 +888,145 @@ describe('session()', function(){
       .get('/foo')
       .expect(shouldSetCookieToValue('connect.sid', 's%3A%2Ffoo.paEKBtAHbV5s1IB8B2zPnzAgYmmnRPIqObW4VRYj%2FMQ'))
       .expect(200, done)
+    });
+  });
+
+  describe('allowUnsigned option', function() {
+    it('should default to false', function(done) {
+      var sessionId;
+      var accessToken;
+      var secret = 's3cret';
+
+      var serverOptions = {secret: secret, rolling: true};
+
+      var server = createServer(serverOptions, function(req, res) {
+        if (accessToken) return res.end();
+        accessToken = session.getCookieValue(req, 'connect.sid', secret);
+        sessionId = req.sessionID;
+        return res.end();
+      });
+
+      request(server)
+        .get('/')
+        .expect(200, function(err) {
+          if (err) done(err);
+          request(server)
+            .get('/')
+            .set('Cookie', ['connect.sid=' + sessionId])
+            .expect(shouldSetCookieToDifferentSessionId(sessionId))
+            .expect(200, done);
+        });
+    });
+
+    it('should allow unsigned values when set to true', function(done) {
+      var sessionId;
+      var accessToken;
+      var secret = 's3cret';
+
+      var serverOptions = {secret: secret, rolling: true, allowUnsigned: true};
+
+      var server = createServer(serverOptions, function(req, res) {
+        if (accessToken) return res.end();
+        accessToken = session.getCookieValue(req, 'connect.sid', secret);
+        sessionId = req.sessionID;
+        return res.end();
+      });
+
+      request(server)
+        .get('/')
+        .expect(200, function(err) {
+          if (err) done(err);
+          request(server)
+            .get('/')
+            .set('Cookie', ['connect.sid=' + sessionId])
+            .expect(shouldSetCookieToValue('connect.sid', accessToken))
+            .expect(200, done);
+        });
+    });
+  });
+
+  describe('alternateTokenValue option', function() {
+    it('should reject non-function values', function() {
+      assert.throws(
+        session.bind(null, {alternateTokenValue: 'bogus!'}),
+        /alternateTokenValue.*must/
+      );
+    });
+
+    it('should ignore alternate locations and return undefined when not set', function(done) {
+      var sessionId;
+      var accessToken;
+      var secret = 's3cret';
+
+      var serverOptions = {secret: secret, rolling: true};
+
+      var server = createServer(serverOptions, function(req, res) {
+        if (accessToken) return res.end();
+        accessToken = session.getCookieValue(req, 'connect.sid', secret);
+        sessionId = req.sessionID;
+        return res.end();
+      });
+
+      request(server)
+        .get('/')
+        .expect(200, function(err) {
+          if (err) done(err);
+          request(server)
+            .get('/?accessToken=' + accessToken)
+            .expect(shouldSetCookieToDifferentSessionId(sessionId))
+            .expect(200, done);
+        });
+    });
+
+    it('should get cookie value from supplied function when available', function(done) {
+      var accessToken;
+      var secret = 's3cret';
+
+      var alternateTokenValue = function(req) {
+        var q = decodeURIComponent(req.url.split('?')[1]);
+        return q ? q.split('accessToken=')[1] : undefined;
+      };
+
+      var serverOptions = {secret: secret, rolling: true, alternateTokenValue: alternateTokenValue};
+
+      var server = createServer(serverOptions, function(req, res) {
+        if (accessToken) return res.end();
+        accessToken = session.getCookieValue(req, 'connect.sid', secret);
+        return res.end();
+      });
+
+      request(server)
+        .get('/')
+        .expect(200, function(err) {
+          if (err) done(err);
+          request(server)
+            .get('/?accessToken=' + accessToken)
+            .expect(shouldSetCookieToValue('connect.sid', accessToken))
+            .expect(200, done);
+        });
+    });
+  });
+
+  describe('skipCookie option', function() {
+    it('should reject non-function values', function() {
+      assert.throws(
+        session.bind(null, {skipCookie: 'bogus!'}),
+        /skipCookie.*must/
+      );
+    });
+
+    it('should default to false', function(done) {
+      request(createServer())
+        .get('/')
+        .expect(shouldSetCookie('connect.sid'))
+        .expect(200, done);
+    });
+
+    it('should not set cookie when function returns true', function(done) {
+      request(createServer({skipCookie: function() { return true }}))
+        .get('/')
+        .expect(shouldNotHaveHeader('Set-Cookie'))
+        .expect(200, done);
     });
   });
 
@@ -1396,7 +1562,7 @@ describe('session()', function(){
   describe('res.end patch', function () {
     it('should correctly handle res.end/res.write patched prior', function (done) {
       function setup (req, res) {
-        writePatch(res)
+        utils.writePatch(res)
       }
 
       function respond (req, res) {
@@ -1412,7 +1578,7 @@ describe('session()', function(){
 
     it('should correctly handle res.end/res.write patched after', function (done) {
       function respond (req, res) {
-        writePatch(res)
+        utils.writePatch(res)
         req.session.hit = true
         res.write('hello, ')
         res.end('world')
@@ -1824,6 +1990,55 @@ describe('session()', function(){
         })
       })
 
+      describe('.originalMaxAge', function () {
+        it('should equal original maxAge', function (done) {
+          var server = createServer({ cookie: { maxAge: 2000 } }, function (req, res) {
+            res.end(JSON.stringify(req.session.cookie.originalMaxAge))
+          })
+
+          request(server)
+            .get('/')
+            .expect(200, '2000', done)
+        })
+
+        it('should equal original maxAge for all requests', function (done) {
+          var server = createServer({ cookie: { maxAge: 2000 } }, function (req, res) {
+            res.end(JSON.stringify(req.session.cookie.originalMaxAge))
+          })
+
+          request(server)
+            .get('/')
+            .expect(200, '2000', function (err, res) {
+              if (err) return done(err)
+              setTimeout(function () {
+                request(server)
+                  .get('/')
+                  .set('Cookie', cookie(res))
+                  .expect(200, '2000', done)
+              }, 100)
+            })
+        })
+
+        it('should equal original maxAge for all requests', function (done) {
+          var store = new SmartStore()
+          var server = createServer({ cookie: { maxAge: 2000 }, store: store }, function (req, res) {
+            res.end(JSON.stringify(req.session.cookie.originalMaxAge))
+          })
+
+          request(server)
+            .get('/')
+            .expect(200, '2000', function (err, res) {
+              if (err) return done(err)
+              setTimeout(function () {
+                request(server)
+                  .get('/')
+                  .set('Cookie', cookie(res))
+                  .expect(200, '2000', done)
+              }, 100)
+            })
+        })
+      })
+
       describe('.secure', function(){
         var app
 
@@ -2185,7 +2400,7 @@ function end(req, res) {
 
 function expires (res) {
   var header = cookie(res)
-  return header && parseSetCookie(header).expires
+  return header && utils.parseSetCookie(header).expires
 }
 
 function mountAt (path) {
@@ -2195,25 +2410,6 @@ function mountAt (path) {
       req.url = req.url.slice(path.length)
     }
   }
-}
-
-function parseSetCookie (header) {
-  var match
-  var pairs = []
-  var pattern = /\s*([^=;]+)(?:=([^;]*);?|;|$)/g
-
-  while ((match = pattern.exec(header))) {
-    pairs.push({ name: match[1], value: match[2] })
-  }
-
-  var cookie = pairs.shift()
-
-  for (var i = 0; i < pairs.length; i++) {
-    match = pairs[i]
-    cookie[match.name.toLowerCase()] = (match.value || true)
-  }
-
-  return cookie
 }
 
 function shouldNotHaveHeader(header) {
@@ -2239,7 +2435,7 @@ function shouldNotSetSessionInStore(store) {
 function shouldSetCookie (name) {
   return function (res) {
     var header = cookie(res)
-    var data = header && parseSetCookie(header)
+    var data = header && utils.parseSetCookie(header)
     assert.ok(header, 'should have a cookie header')
     assert.strictEqual(data.name, name, 'should set cookie ' + name)
   }
@@ -2254,7 +2450,7 @@ function shouldSetCookieToDifferentSessionId (id) {
 function shouldSetCookieToExpireIn (name, delta) {
   return function (res) {
     var header = cookie(res)
-    var data = header && parseSetCookie(header)
+    var data = header && utils.parseSetCookie(header)
     assert.ok(header, 'should have a cookie header')
     assert.strictEqual(data.name, name, 'should set cookie ' + name)
     assert.ok(('expires' in data), 'should set cookie with attribute Expires')
@@ -2266,7 +2462,7 @@ function shouldSetCookieToExpireIn (name, delta) {
 function shouldSetCookieToValue (name, val) {
   return function (res) {
     var header = cookie(res)
-    var data = header && parseSetCookie(header)
+    var data = header && utils.parseSetCookie(header)
     assert.ok(header, 'should have a cookie header')
     assert.strictEqual(data.name, name, 'should set cookie ' + name)
     assert.strictEqual(data.value, val, 'should set cookie ' + name + ' to ' + val)
@@ -2276,7 +2472,7 @@ function shouldSetCookieToValue (name, val) {
 function shouldSetCookieWithAttribute (name, attrib) {
   return function (res) {
     var header = cookie(res)
-    var data = header && parseSetCookie(header)
+    var data = header && utils.parseSetCookie(header)
     assert.ok(header, 'should have a cookie header')
     assert.strictEqual(data.name, name, 'should set cookie ' + name)
     assert.ok((attrib.toLowerCase() in data), 'should set cookie with attribute ' + attrib)
@@ -2286,7 +2482,7 @@ function shouldSetCookieWithAttribute (name, attrib) {
 function shouldSetCookieWithAttributeAndValue (name, attrib, value) {
   return function (res) {
     var header = cookie(res)
-    var data = header && parseSetCookie(header)
+    var data = header && utils.parseSetCookie(header)
     assert.ok(header, 'should have a cookie header')
     assert.strictEqual(data.name, name, 'should set cookie ' + name)
     assert.ok((attrib.toLowerCase() in data), 'should set cookie with attribute ' + attrib)
@@ -2297,7 +2493,7 @@ function shouldSetCookieWithAttributeAndValue (name, attrib, value) {
 function shouldSetCookieWithoutAttribute (name, attrib) {
   return function (res) {
     var header = cookie(res)
-    var data = header && parseSetCookie(header)
+    var data = header && utils.parseSetCookie(header)
     assert.ok(header, 'should have a cookie header')
     assert.strictEqual(data.name, name, 'should set cookie ' + name)
     assert.ok(!(attrib.toLowerCase() in data), 'should set cookie without attribute ' + attrib)
@@ -2320,48 +2516,8 @@ function shouldSetSessionInStore(store) {
 
 function sid (res) {
   var header = cookie(res)
-  var data = header && parseSetCookie(header)
+  var data = header && utils.parseSetCookie(header)
   var value = data && unescape(data.value)
   var sid = value && value.substring(2, value.indexOf('.'))
   return sid || undefined
 }
-
-function writePatch (res) {
-  var _end = res.end
-  var _write = res.write
-  var ended = false
-
-  res.end = function end() {
-    ended = true
-    return _end.apply(this, arguments)
-  }
-
-  res.write = function write() {
-    if (ended) {
-      throw new Error('write after end')
-    }
-
-    return _write.apply(this, arguments)
-  }
-}
-
-function SyncStore () {
-  session.Store.call(this)
-  this.sessions = Object.create(null)
-}
-
-util.inherits(SyncStore, session.Store)
-
-SyncStore.prototype.destroy = function destroy(sid, callback) {
-  delete this.sessions[sid];
-  callback();
-};
-
-SyncStore.prototype.get = function get(sid, callback) {
-  callback(null, JSON.parse(this.sessions[sid]));
-};
-
-SyncStore.prototype.set = function set(sid, sess, callback) {
-  this.sessions[sid] = JSON.stringify(sess);
-  callback();
-};